Replace multi-agent translation test with program creation example
